// src/types/likeds.ts
export type LikedsReceive = {
  user_id: string
  id: string
  age: number | null
  business: string | null
  gender: string
  hobby: string | null
  introduction_summary: string | null
  main_photo: string | null
  message: string | null
  nickname: string
  prefecture: string
  is_online: boolean
}